+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.3] - 2025-08-21
 ### Added
+- **Settings Preset Manager:**
+  - Save / Delete / Load presets,
+  - Export settings state to .json.
+  - Import .json to preset.
+### Changed
+- **BBMM:** Changed "Preset Manager" button to "BBMM" on manage modules window, opens window to choose Module or Settings preset manager
+
+## [0.0.3] - 2025-08-21
+### Added
 - **Module Preset Manager:** Added "Preset Manager" button to manage modules window.
 - **Module Preset Manager:** Added overwrite protection for presets. 
   - If you try and save/import a preset with an existing name, it will prompt overwrite/rename.
@@ -18,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.1] - 2025-08-20
 ### Added
 - **Initial Upload:** Initial upload.
-- **Preset Manager:**
+- **Module Preset Manager:**
   - Save / Delete / Load presets,
   - Export module state to .json.
   - Import .json to preset.